Refactor codes for using common codes
