+++ global-inject-demo/global-inject-demo.vcxproj.filters
@@ -27,7 +27,7 @@
     <ClCompile Include="engine_control.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\shared\common\functions.cpp">
+    <ClCompile Include="..\shared\common_functions.cpp">
       <Filter>common</Filter>
     </ClCompile>
   </ItemGroup>
@@ -41,7 +41,7 @@
     <ClInclude Include="functions.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\shared\common\functions.h">
+    <ClInclude Include="..\shared\common_functions.h">
       <Filter>common</Filter>
     </ClInclude>
   </ItemGroup>
